+++ WheresWally/Struct.h
@@ -3,5 +3,5 @@
 struct CoOrd {
 	int x;
 	int y;
-	int dist;
+	double dist;
 };
